Move InvocationContext and InvocationEventHandler to com.palantir.tritium.api
