+++ tritium-api/src/main/java/com/palantir/tritium/event/InvocationEventHandler.java
@@ -16,54 +16,14 @@
 
 package com.palantir.tritium.event;
 
-import java.lang.reflect.Method;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-
 /**
  * Interface for handing invocation events.
+ * Back-compat bridge type, will be removed in future major version bump.
  *
  * @see java.lang.reflect.InvocationHandler
  * @param <C> invocation context
+ * @deprecated use {@link com.palantir.tritium.api.event.InvocationEventHandler}
  */
-public interface InvocationEventHandler<C extends InvocationContext> {
-
-    /**
-     * Returns true if this event handler instance is enabled, otherwise false.
-     *
-     * @return true if handler is enabled
-     */
-    boolean isEnabled();
-
-    /**
-     * Invoked before invoking the method on the instance.
-     *
-     * @param instance the instance that the method was invoked on.
-     * @param method the {@code Method} corresponding to the interface method invoked on the instance.
-     * @param args an array of objects containing the values of the arguments passed in the method invocation on the
-     *     instance, or empty array if interface method takes no arguments. Arguments of primitive types are wrapped in
-     *     instances of the appropriate primitive wrapper class, such as {@code java.lang.Integer} or
-     *     {@code java.lang.Boolean}.
-     * @return the current invocation context. Null values are not recommended but are supported
-     */
-    C preInvocation(@Nonnull Object instance, @Nonnull Method method, @Nonnull Object[] args);
-
-    /**
-     * Invoked with the result of the invocation when it is successful.
-     *
-     * @param context the current invocation context or null if preInvocation returned null, or threw an exception.
-     * @param result the return value from invocation, or null if {@link Void}.
-     */
-    void onSuccess(@Nullable InvocationContext context, @Nullable Object result);
-
-    /**
-     * Invoked when an invocation fails.
-     *
-     * <p>If the invocation throws an {@link java.lang.reflect.InvocationTargetException}, then the cause is passed to
-     * this method. Any other thrown object is passed unaltered.
-     *
-     * @param context the current invocation context or null if preInvocation returned null, or threw an exception.
-     * @param cause the throwable which caused the failure.
-     */
-    void onFailure(@Nullable InvocationContext context, @Nonnull Throwable cause);
-}
+@Deprecated
+public interface InvocationEventHandler<C extends InvocationContext>
+        extends com.palantir.tritium.api.event.InvocationEventHandler<C> {}
