+++ tritium-core/src/main/java/com/palantir/tritium/event/AbstractInvocationEventHandler.java
@@ -19,6 +19,9 @@ package com.palantir.tritium.event;
 import static com.palantir.logsafe.Preconditions.checkNotNull;
 
 import com.palantir.logsafe.SafeArg;
+import com.palantir.tritium.api.event.InvocationContext;
+import com.palantir.tritium.api.event.InvocationEventHandler;
+import java.util.function.BooleanSupplier;
 import javax.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,22 +37,22 @@ public abstract class AbstractInvocationEventHandler<C extends InvocationContext
 
     private static final Object[] NO_ARGS = {};
 
-    private final java.util.function.BooleanSupplier isEnabledSupplier;
+    private final BooleanSupplier isEnabledSupplier;
 
     /** Always enabled instrumentation handler. */
     protected AbstractInvocationEventHandler() {
-        this((java.util.function.BooleanSupplier) () -> true);
+        this((BooleanSupplier) () -> true);
     }
 
     /**
      * Bridge for backward compatibility.
      *
-     * @deprecated Use {@link #AbstractInvocationEventHandler(java.util.function.BooleanSupplier)}
+     * @deprecated Use {@link #AbstractInvocationEventHandler(BooleanSupplier)}
      */
     @Deprecated
     @SuppressWarnings("FunctionalInterfaceClash") // back compat
     protected AbstractInvocationEventHandler(com.palantir.tritium.api.functions.BooleanSupplier isEnabledSupplier) {
-        this((java.util.function.BooleanSupplier) isEnabledSupplier);
+        this((BooleanSupplier) isEnabledSupplier);
     }
 
     /**
@@ -58,7 +61,7 @@ public abstract class AbstractInvocationEventHandler<C extends InvocationContext
      * @param isEnabledSupplier enabled supplier
      */
     @SuppressWarnings("FunctionalInterfaceClash")
-    protected AbstractInvocationEventHandler(java.util.function.BooleanSupplier isEnabledSupplier) {
+    protected AbstractInvocationEventHandler(BooleanSupplier isEnabledSupplier) {
         this.isEnabledSupplier = checkNotNull(isEnabledSupplier, "isEnabledSupplier");
     }
 
@@ -95,6 +98,7 @@ public abstract class AbstractInvocationEventHandler<C extends InvocationContext
      * @param clazz instrumentation handler class
      * @return false if "instrument.fully.qualified.class.Name" is set to "false", otherwise true
      */
+    @SuppressWarnings("deprecation") // back-compat return type for now
     protected static com.palantir.tritium.api.functions.BooleanSupplier getSystemPropertySupplier(
             Class<? extends InvocationEventHandler<InvocationContext>> clazz) {
         checkNotNull(clazz, "clazz");
