+++ tritium-core/src/main/java/com/palantir/tritium/event/NoOpInvocationEventHandler.java
@@ -16,6 +16,8 @@
 
 package com.palantir.tritium.event;
 
+import com.palantir.tritium.api.event.InvocationContext;
+import com.palantir.tritium.api.event.InvocationEventHandler;
 import java.lang.reflect.Method;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
