+++ tritium-core/src/test/java/com/palantir/tritium/event/DefaultInvocationContextTest.java
@@ -18,6 +18,7 @@ package com.palantir.tritium.event;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.palantir.tritium.api.event.InvocationContext;
 import org.junit.jupiter.api.Test;
 
 final class DefaultInvocationContextTest {
