+++ tritium-jmh/src/jmh/java/com/palantir/tritium/microbenchmarks/BlackholeInvocationEventHandler.java
@@ -17,9 +17,9 @@
 package com.palantir.tritium.microbenchmarks;
 
 import com.palantir.logsafe.Preconditions;
+import com.palantir.tritium.api.event.InvocationContext;
+import com.palantir.tritium.api.event.InvocationEventHandler;
 import com.palantir.tritium.event.DefaultInvocationContext;
-import com.palantir.tritium.event.InvocationContext;
-import com.palantir.tritium.event.InvocationEventHandler;
 import java.lang.reflect.Method;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
