+++ tritium-lib/src/main/java/com/palantir/tritium/proxy/InstrumentationProxy.java
@@ -17,8 +17,8 @@
 package com.palantir.tritium.proxy;
 
 import com.palantir.tritium.api.event.InstrumentationFilter;
-import com.palantir.tritium.event.InvocationContext;
-import com.palantir.tritium.event.InvocationEventHandler;
+import com.palantir.tritium.api.event.InvocationContext;
+import com.palantir.tritium.api.event.InvocationEventHandler;
 import java.util.List;
 
 class InstrumentationProxy<T> extends InvocationEventProxy {
