+++ tritium-lib/src/test/java/com/palantir/tritium/proxy/InstrumentationTest.java
@@ -40,11 +40,11 @@ import com.google.common.util.concurrent.Runnables;
 import com.palantir.logsafe.exceptions.SafeRuntimeException;
 import com.palantir.tritium.Tagged;
 import com.palantir.tritium.api.event.InstrumentationFilter;
+import com.palantir.tritium.api.event.InvocationContext;
+import com.palantir.tritium.api.event.InvocationEventHandler;
 import com.palantir.tritium.event.DefaultInvocationContext;
 import com.palantir.tritium.event.InstrumentationFilters;
 import com.palantir.tritium.event.InstrumentationProperties;
-import com.palantir.tritium.event.InvocationContext;
-import com.palantir.tritium.event.InvocationEventHandler;
 import com.palantir.tritium.event.log.LoggingInvocationEventHandler;
 import com.palantir.tritium.event.metrics.MetricsInvocationEventHandler;
 import com.palantir.tritium.event.metrics.TaggedMetricsServiceInvocationEventHandler;
