+++ tritium-tracing/src/test/java/com/palantir/tritium/tracing/TracingInvocationEventHandlerTest.java
@@ -28,8 +28,8 @@ import com.palantir.tracing.AsyncSlf4jSpanObserver;
 import com.palantir.tracing.Tracer;
 import com.palantir.tracing.api.Span;
 import com.palantir.tracing.api.SpanObserver;
-import com.palantir.tritium.event.InvocationContext;
-import com.palantir.tritium.event.InvocationEventHandler;
+import com.palantir.tritium.api.event.InvocationContext;
+import com.palantir.tritium.api.event.InvocationEventHandler;
 import com.palantir.tritium.test.TestImplementation;
 import com.palantir.tritium.test.TestInterface;
 import java.lang.reflect.Method;
@@ -152,7 +152,7 @@ public class TracingInvocationEventHandlerTest {
 
     @Test
     public void testSystemPropertySupplier_Handler_Enabled() {
-        assertThat(TracingInvocationEventHandler.getEnabledSupplier("test").asBoolean())
+        assertThat(TracingInvocationEventHandler.getEnabledSupplier("test").getAsBoolean())
                 .isTrue();
     }
 }
